CIProvider logic moved to internal-api module
