+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/ci/AzurePipelinesInfo.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.ci;
+package datadog.trace.bootstrap.instrumentation.api.ci;
 
 class AzurePipelinesInfo extends CIProviderInfo {
