+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/ci/GitLabInfo.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.ci;
+package datadog.trace.bootstrap.instrumentation.api.ci;
 
 class GitLabInfo extends CIProviderInfo {
