+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/api/ci/AzurePipelinesInfoTest.groovy
@@ -1,20 +1,21 @@
-package datadog.trace.api.ci
+package datadog.trace.bootstrap.instrumentation.api.ci
 
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_BUILD_BUILDID
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_BUILD_REPOSITORY_URI
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_BUILD_SOURCEBRANCH
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_BUILD_SOURCEVERSION
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_PIPELINE_NAME
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_PROVIDER_NAME
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_JOBID
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCEBRANCH
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCECOMMITID
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCEREPOSITORYURI
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TASKINSTANCEID
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TEAMFOUNDATIONSERVERURI
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TEAMPROJECT
-import static datadog.trace.api.ci.AzurePipelinesInfo.AZURE_WORKSPACE_PATH
+
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_BUILD_BUILDID
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_BUILD_REPOSITORY_URI
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_BUILD_SOURCEBRANCH
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_BUILD_SOURCEVERSION
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_PIPELINE_NAME
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_PROVIDER_NAME
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_JOBID
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCEBRANCH
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCECOMMITID
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_PULLREQUEST_SOURCEREPOSITORYURI
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TASKINSTANCEID
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TEAMFOUNDATIONSERVERURI
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_SYSTEM_TEAMPROJECT
+import static datadog.trace.bootstrap.instrumentation.api.ci.AzurePipelinesInfo.AZURE_WORKSPACE_PATH
 
 class AzurePipelinesInfoTest extends CIProviderInfoTest {
