+++ internal-api/src/test/groovy/datadog/trace/bootstrap/instrumentation/api/ci/NoopCIInfoTest.groovy
@@ -1,4 +1,4 @@
-package datadog.trace.api.ci
+package datadog.trace.bootstrap.instrumentation.api.ci
 
 class NoopCIInfoTest extends CIProviderInfoTest {
